+++ src/actions/args.rs
@@ -1,9 +1,11 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 // Copyright (C) 2024 Tobias Hunger <tobias.hunger@gmail.com>
 
+// cSpell: ignore concatcp dotdotdot starstar
+
 use std::cell::RefCell;
 use std::ffi::{OsStr, OsString};
-use std::path::PathBuf;
+use std::path::{Path, PathBuf};
 
 use crate::actions::inputs;
 
@@ -144,13 +146,13 @@ fn split_arg(arg: &str) -> Vec<String> {
 async fn input_arg(
     arg: &str,
     inputs: inputs::InputQuery,
+    root_directory: &Path,
     input_filters: &inputs::InputFilters,
 ) -> crate::Result<Option<(Vec<PathBuf>, bool)>> {
     static EMPTY: Vec<glob::Pattern> = vec![];
 
     if arg.starts_with("{{") && arg.ends_with("}}") {
         let input_name = &arg[2..(arg.len() - 2)];
-        tracing::debug!("{arg} is an input argument with name: {input_name}");
         let (input_name, is_array) = if input_name.ends_with("...") {
             (&input_name[0..(input_name.len() - 3)], true)
         } else {
@@ -158,6 +160,11 @@ async fn input_arg(
         };
 
         let current_filters = input_filters.get(input_name).unwrap_or(&EMPTY);
+        let match_options = {
+            let mut opt = glob::MatchOptions::new();
+            opt.require_literal_separator = true;
+            opt
+        };
 
         let paths = inputs
             .inputs(input_name.to_string())
@@ -165,13 +172,16 @@ async fn input_arg(
             .map_err(|e| crate::Error::new_input_generator(input_name.to_string(), e.to_string()))?
             .into_iter()
             .filter(|p| {
-                current_filters.is_empty() || current_filters.iter().any(|f| f.matches_path(p))
+                let Ok(rel_path) = p.strip_prefix(root_directory) else {
+                    return false;
+                };
+                current_filters.is_empty()
+                    || current_filters.iter().any(|f| f.matches_path_with(rel_path, match_options.clone()))
             })
             .collect();
 
         Ok(Some((paths, is_array)))
     } else {
-        tracing::trace!("{arg} is no input argument");
         Ok(None)
     }
 }
@@ -180,6 +190,7 @@ async fn input_arg(
 pub(crate) async fn parse_arg(
     arg: &str,
     inputs: inputs::InputQuery,
+    root_directory: &Path,
     input_filters: &inputs::InputFilters,
 ) -> crate::Result<Option<Vec<Arg>>> {
     let argument_parts = split_arg(arg);
@@ -188,7 +199,9 @@ pub(crate) async fn parse_arg(
 
     if argument_parts.len() == 1 {
         let arg = &argument_parts[0];
-        if let Some((paths, is_array)) = input_arg(arg, inputs, input_filters).await? {
+        if let Some((paths, is_array)) =
+            input_arg(arg, inputs, root_directory, input_filters).await?
+        {
             if paths.is_empty() {
                 return Ok(None);
             }
@@ -211,7 +224,9 @@ pub(crate) async fn parse_arg(
         let mut extended_arg = vec![String::new()];
 
         for p in &argument_parts {
-            if let Some((paths, is_array)) = input_arg(p, inputs.clone(), input_filters).await? {
+            if let Some((paths, is_array)) =
+                input_arg(p, inputs.clone(), root_directory, input_filters).await?
+            {
                 if paths.is_empty() {
                     return Ok(None);
                 }
@@ -254,12 +269,13 @@ pub(crate) async fn parse_arg(
 pub(crate) async fn parse_args(
     args: &[String],
     inputs: inputs::InputQuery,
+    root_directory: &Path,
     input_filters: &inputs::InputFilters,
 ) -> crate::Result<Option<Args>> {
     let mut parsed_args = Vec::with_capacity(args.len() - 1);
 
     for a in args.iter().skip(1) {
-        let filtered_args = parse_arg(a, inputs.clone(), input_filters).await?;
+        let filtered_args = parse_arg(a, inputs.clone(), root_directory, input_filters).await?;
         if let Some(filtered_args) = filtered_args {
             parsed_args.extend_from_slice(&filtered_args);
         } else {
@@ -272,6 +288,8 @@ pub(crate) async fn parse_args(
 
 #[cfg(test)]
 mod tests {
+    use std::collections::HashMap;
+
     use super::*;
 
     #[test]
@@ -296,5 +314,156 @@ mod tests {
             ],
             split_arg("test{{files...}}foobar")
         );
+    }
+
+    const ROOT_DIR: &str = if cfg!(windows) {
+        "C:\\51bb3d94"
+    } else {
+        "/tmp/51bb3d94"
+    };
+
+    const PATH_0: &str = &const_format::concatcp!(ROOT_DIR, std::path::MAIN_SEPARATOR, "README.md");
+    const PATH_1: &str = &const_format::concatcp!(ROOT_DIR, std::path::MAIN_SEPARATOR, "main.rs");
+    const PATH_2: &str =
+        &const_format::concatcp!(ROOT_DIR, std::path::MAIN_SEPARATOR, "Cargo.toml");
+    const PATH_3: &str = &const_format::concatcp!(
+        ROOT_DIR,
+        std::path::MAIN_SEPARATOR,
+        "docs",
+        std::path::MAIN_SEPARATOR,
+        "doc.md"
+    );
+    const PATH_4: &str = &const_format::concatcp!(
+        ROOT_DIR,
+        std::path::MAIN_SEPARATOR,
+        "docs",
+        std::path::MAIN_SEPARATOR,
+        "sections",
+        std::path::MAIN_SEPARATOR,
+        "one.md"
+    );
+
+    async fn test_input_arg(
+        arg: &str,
+        filters: &[&str],
+    ) -> crate::Result<Option<(Vec<PathBuf>, bool)>> {
+        let input_cache = inputs::setup_input_cache(
+            PathBuf::from(ROOT_DIR),
+            vec![
+                PathBuf::from(PATH_0),
+                PathBuf::from(PATH_1),
+                PathBuf::from(PATH_2),
+                PathBuf::from(PATH_3),
+                PathBuf::from(PATH_4),
+            ],
+        );
+
+        let filter = HashMap::from([(
+            "files".to_string(),
+            filters
+                .iter()
+                .map(|f| glob::Pattern::new(f).unwrap())
+                .collect(),
+        )]);
+
+        let root_directory = PathBuf::from(ROOT_DIR);
+
+        let result = input_arg(arg, input_cache.query(), &root_directory, &filter).await;
+        result
+    }
+
+    #[tokio::test]
+    async fn test_input_arg_none() {
+        let result = test_input_arg("foo", &[]).await.unwrap();
+
+        assert!(result.is_none());
+    }
+
+    #[tokio::test]
+    async fn test_input_arg_files() {
+        let (paths, is_array) = test_input_arg("{{files}}", &[]).await.unwrap().unwrap();
+
+        assert_eq!(is_array, false);
+        assert_eq!(paths.len(), 5);
+        assert_eq!(paths[0].to_string_lossy(), PATH_0);
+        assert_eq!(paths[1].to_string_lossy(), PATH_1);
+        assert_eq!(paths[2].to_string_lossy(), PATH_2);
+        assert_eq!(paths[3].to_string_lossy(), PATH_3);
+        assert_eq!(paths[4].to_string_lossy(), PATH_4);
+    }
+
+    #[tokio::test]
+    async fn test_input_arg_files_dotdotdot() {
+        let (paths, is_array) = test_input_arg("{{files...}}", &[]).await.unwrap().unwrap();
+
+        assert_eq!(is_array, true);
+        assert_eq!(paths.len(), 5);
+        assert_eq!(paths[0].to_string_lossy(), PATH_0);
+        assert_eq!(paths[1].to_string_lossy(), PATH_1);
+        assert_eq!(paths[2].to_string_lossy(), PATH_2);
+        assert_eq!(paths[3].to_string_lossy(), PATH_3);
+        assert_eq!(paths[4].to_string_lossy(), PATH_4);
+    }
+
+    #[tokio::test]
+    async fn test_input_arg_files_filter_starstar_star_md() {
+        let (paths, is_array) = test_input_arg("{{files...}}", &["**/*.md"])
+            .await
+            .unwrap()
+            .unwrap();
+
+        assert_eq!(is_array, true);
+        assert_eq!(paths.len(), 3);
+        assert_eq!(paths[0].to_string_lossy(), PATH_0);
+        assert_eq!(paths[1].to_string_lossy(), PATH_3);
+        assert_eq!(paths[2].to_string_lossy(), PATH_4);
+    }
+
+    #[tokio::test]
+    async fn test_input_arg_files_filter_star_md() {
+        let (paths, is_array) = test_input_arg("{{files...}}", &["*.md"])
+            .await
+            .unwrap()
+            .unwrap();
+
+        assert_eq!(is_array, true);
+        assert_eq!(paths.len(), 1);
+        assert_eq!(paths[0].to_string_lossy(), PATH_0);
+    }
+
+    #[tokio::test]
+    async fn test_input_arg_files_filter_docs_md() {
+        let (paths, is_array) = test_input_arg("{{files...}}", &["docs/*.md"])
+            .await
+            .unwrap()
+            .unwrap();
+
+        assert_eq!(is_array, true);
+        assert_eq!(paths.len(), 1);
+        assert_eq!(paths[0].to_string_lossy(), PATH_3);
+    }
+
+    #[tokio::test]
+    async fn test_input_arg_files_filter_docs_starstar_md() {
+        let (paths, is_array) = test_input_arg("{{files...}}", &["docs/**/*.md"])
+            .await
+            .unwrap()
+            .unwrap();
+
+        assert_eq!(is_array, true);
+        assert_eq!(paths.len(), 2);
+        assert_eq!(paths[0].to_string_lossy(), PATH_3);
+        assert_eq!(paths[1].to_string_lossy(), PATH_4);
+    }
+
+    #[tokio::test]
+    async fn test_input_arg_files_filter_github_actions() {
+        let (paths, is_array) = test_input_arg("{{files...}}", &[".github/**/*.yaml", ".github/**/*.yml"])
+            .await
+            .unwrap()
+            .unwrap();
+
+        assert_eq!(is_array, true);
+        assert_eq!(paths.len(), 0);
     }
 }